fix code (separate content show)
