+++ app/views/contents/_editor.html.slim
@@ -25,7 +25,7 @@
           - if content.errors.any?
             = model_errors(content)
           = f.hidden_field :lock_version
-          = f.text_area :body, class: 'tinymce', id: "form-tinymce"
+          = f.text_area :body, class: 'tinymce'
 
           - if can? :update, content
             = link_to t('.ready_for_review'), change_status_protocol_content_path(protocol, content, 'content[status]': 'under_review'),
